Ensemble Techniques Project- Banking Project 1
This is a Banking Domain Project. The data is related with direct marketing campaigns of a Portuguese banking institution. The marketing campaigns were based on phone calls. Often, more than one contact to the same client was required, in order to access if the product (bank term deposit) would be ('yes') or not ('no') subscribed. 
The classification goal is to predict if the client will subscribe (yes/no) a term deposit . There are many categorical and numerical features . The zip file attached contains all plots , graphs and numerical calculations done in this project . People  who wants to learn about the practical  usage of different classification models in Machine Learning can refer to this project